requirements.txt
# Python dependencies for AutoML-Insight

# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0

# Explainability
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dimensionality reduction
umap-learn>=0.5.3

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Web app
streamlit>=1.25.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0

# PDF generation
reportlab>=4.0.0
weasyprint>=59.0

# Hyperparameter tuning
optuna>=3.3.0

# Utilities
joblib>=1.3.0
psutil>=5.9.0

# Remote execution
requests>=2.31.0
pyngrok>=5.1.0

# AI/LLM Integration
python-dotenv>=1.0.0
groq>=0.4.0
openai>=1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-groq>=0.0.1

# Development tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
